+++ src/main/java/lotto/viewHandler/validator/Validator.java
@@ -1,4 +1,4 @@
-package lotto.viewHandler;
+package lotto.viewHandler.validator;
 
 public interface Validator<T1, T2> {
     T1 validate(T2 input);
