+++ src/main/java/lotto/viewHandler/validator/validatorImpl/LottoPurchaseUnitValidator.java
@@ -1,7 +1,7 @@
-package lotto.viewHandler.validator;
+package lotto.viewHandler.validator.validatorImpl;
 
 import lotto.viewHandler.exception.NotUnitPurchaseMoney;
-import lotto.viewHandler.Validator;
+import lotto.viewHandler.validator.Validator;
 
 import static lotto.domain.PurchaseLottos.LOTTO_UNIT;
 import static lotto.viewHandler.exception.MyExceptionConstant.NOT_MONEY_UNIT;
