+++ src/test/java/lotto/viewHandler/validator/LottoNumberSplitTest.java
@@ -1,13 +1,13 @@
 package lotto.viewHandler.validator;
 
 import lotto.viewHandler.exception.NotLottoNumberSize;
+import lotto.viewHandler.validator.validatorImpl.LottoNumberSplit;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
 
 class LottoNumberSplitTest {
     private final LottoNumberSplit lottoNumberSplit;
